Split components into files.
